% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{check_deployment_models}
\alias{check_deployment_models}
\title{Check deployment calibration model diagnostic plots}
\usage{
check_deployment_models(package)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{\link[camtraptor]{read_camtrap_dp}}.}
}
\value{
The original package with logical column \code{useDeployment}`
 added to deployments and observations data.
}
\description{
Displays deployment calibration model diagnostic plots and allows
users to record interactively whether each deployment is reliable.
}
\examples{
  pkg <- camtraptor::read_camtrap_dp("./data/datapackage.json")
  pkg_checked <- check_deployment_models(pkg)
}
