% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{correct_time}
\alias{correct_time}
\title{Correct times for a given deployment in a camera trap datapackage}
\usage{
correct_time(package, deploymentID, wrongTime, rightTime)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{\link[camtraptor]{read_camtrap_dp}}.}

\item{deploymentID}{A character value giving the deployment identifier,
to be matched in package$data$deployments$deploymentID.}

\item{wrongTime}{A character or POSIX reference date-time recorded wrongly
by the camera.}

\item{rightTime}{A character or POSIX value giving the correct date-time
when the reference time was recorded.}
}
\value{
As for    \code{\link[camtraptor]{read_camtrap_dp}}, with all
 date-times corrected by the difference between rightTime and wrongTime.
}
\description{
When a camera trap starts with the wrong time stamp, times
in all datapackage tables can be corrected given a reference time
recorded by the camera, the correct time for this reference time
and the ID of the deployment to correct.
}
\examples{
pkg_corrected <- correct_time(pkg,
                              deploymentID = "0d620d0e-5da8-42e6-bcf2-56c11fb3d08e",
                              wrongTime = "2017-01-01 00:00:00",
                              rightTime = "2018-09-01 11:04:00")
}
