% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{downversion_camtrapDP}
\alias{downversion_camtrapDP}
\title{Downversion Camtrap DP datapackage}
\usage{
downversion_camtrapDP(pkg)
}
\arguments{
\item{pkg}{Camera trap data package list, as returned by
\code{\link{read_camtrapDP}}.}
}
\value{
A duplicate of the original package with the addition of duplicate
 of fields with V0 fieldnames to data.
}
\description{
Downversions a Cmatrap DP datapackage to V0 by duplicating dataframe fields
whose names differ between versions with V0 fields created in addition to
existing V1 fields.
}
\examples{
  \dontrun{pkgV0 <- downversion_camtrapDP(pkgV1)}
}
