% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{filter_camtrap_dp}
\alias{filter_camtrap_dp}
\title{Select a subset of a data package}
\usage{
filter_camtrap_dp(package, depChoice = NULL, start = NULL, end = NULL)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{\link[camtraptor]{read_camtrap_dp}}.}

\item{depChoice}{A logical expression using column names from the
deployments table defining which deployments to keep.}

\item{start, end}{Single character or POSIXct values defining the
time range within which to subset the package.}
}
\value{
As for \code{\link[camtraptor]{read_camtrap_dp}}, with all
 data tables reduced according to the choice criteria.
}
\description{
Discards any observations and media that occur at specified deployments,
or that fall outside the time range defined by start and end. When start
and/end are specified, truncates deployment start/end times, and removes
any deployments that fall entirely outside the subset time range.
}
\examples{
# subset excluding a location and including only data from mid October 2017
  \dontrun{
    pkg <- camtraptor::read_camtrap_dp("./datapackage/datapackage.json")
    }
  data(pkg)
  subpkg <- filter_camtrap_dp(pkg,
                              depChoice = locationName!="S01",
                              start = "2017/10/10",
                              end = "2017/10/20")
}
