% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{fit_detmodel}
\alias{fit_detmodel}
\title{Fit a detection function model}
\usage{
fit_detmodel(
  formula,
  package,
  species = NULL,
  newdata = NULL,
  unit = c("m", "km", "cm", "degree", "radian"),
  ...
)
}
\arguments{
\item{formula}{A two sided formula relating radius or angle data
to covariates.}

\item{package}{Camera trap data package object, as returned by
\code{\link[camtraptor]{read_camtrap_dp}}.}

\item{species}{A character string indicating species subset to analyse;
if NULL runs select_species to get user input; if "all" all data are used.}

\item{newdata}{A dataframe of covariate values at which to predict detection
distance.}

\item{unit}{The units in which to return the result.}

\item{...}{Arguments passed to \code{\link[Distance]{ds}}.}
}
\value{
A \code{ddf} detection function model list, with additional elements:
  \code{edd}, a vector with estimated and standard error effective detection
    distance, or the \code{newdata} dataframe with EDD estimate and se
    columns added;
  \code{proportion_used}, the proportion of the observations used to fit the
    detection function in the case of truncation.
}
\description{
Fits a detection function to a data package and estimates effective
detection distance (EDD).
}
\details{
The type of detection function (line or point) is determined by
  the \code{unit} argument.
}
\examples{
  \dontrun{
    pkg <- camtraptor::read_camtrap_dp("./datapackage/datapackage.json")
    }
  data(pkg)
  \dontrun{
    # With interactive species definition
    radius_model <- fit_detmodel(radius~1, pkg, order=0)
  }
  # With species predefined
  sp <- "Vulpes vulpes"
  radius_model <- fit_detmodel(radius~1, pkg, species=sp, order=0)
  angle_model <- fit_detmodel(angle~1, pkg, species=sp, order=0, unit="degree")
  radius_model$edd
  angle_model$edd
  plot(radius_model, pdf=TRUE)
  plot(angle_model)
}
\seealso{
\code{\link[Distance]{ds}}
}
