% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{get_agouti_url}
\alias{get_agouti_url}
\title{Gets a set Agouti sequences URLs.}
\usage{
get_agouti_url(package, obsChoice)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{\link[camtraptor]{read_camtrap_dp}}.}

\item{obsChoice}{A logical expression using column names from the
observations table defining which observations you want to inspect.}
}
\value{
A dataframe of Agouti URLs.
}
\description{
Obtains web addresses for sequences of selected observations, based
on criteria defined using fields in the observations table.
}
\examples{
  data(pkg)
  get_agouti_url(pkg, speed>1)

}
