% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{get_multiplier}
\alias{get_multiplier}
\title{Get a unit multiplier}
\usage{
get_multiplier(unitIN, unitOUT)
}
\arguments{
\item{unitIN}{A text value giving the unit of the input; must be one of
"cm", "m", "km" for distances, "second", "minute", "hour", "day" for
times, "radian" "degree" for angles, or "n/ha", "n/km2" "n/100km2" for
density.}

\item{unitOUT}{As for \code{unitIN}; must be of the same type (distance,
time or density) as \code{unitIN}.}
}
\value{
A number giving the amount by which to multiply input values
  to arrive a unit-converted output.
}
\description{
Returns a multiplier to convert a value from one unit to another in
one of three types: distance, time and angle.
}
\examples{
  get_multiplier("m", "km")
}
