% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{get_parameter_table}
\alias{get_parameter_table}
\title{Create a parameter table from a set of models}
\usage{
get_parameter_table(
  traprate_data,
  radius_model,
  angle_model,
  speed_model,
  activity_model,
  strata = NULL,
  reps = 999,
  seed = 42
)
}
\arguments{
\item{traprate_data}{A dataframe of trap rate data (counts and effort), as
returned by \code{\link{get_traprate_data}}.}

\item{radius_model}{A detection radius model fitted using \code{\link{fit_detmodel}}}

\item{angle_model}{A detection angle model fitted using \code{\link{fit_detmodel}}}

\item{speed_model}{A speed model fitted using \code{\link{fit_speedmodel}}}

\item{activity_model}{An activity  model fitted using \code{\link{fit_actmodel}}}

\item{strata}{A dataframe of stratum information passed to \code{\link{get_trap_rate}}}

\item{reps}{Number of bootstrap replicates for estimating trap rate error
(see \code{\link{get_trap_rate}})}

\item{seed}{Integer seed passed to \code{get_trap_rate} for reproducible bootstrapping; `NULL` to skip setting a seed. Default 42.}
}
\value{
A dataframe of unit-harmonised parameter estimates with rows:
\itemize{
  \item{\code{radius}: detection radius}
  \item{\code{angle}: detection angle}
  \item{\code{active_speed}: speed while active}
  \item{\code{activity_level}: proportion of time spent active}
  \item{\code{overall_speed}: long-term average speed (day range) - the product
    of \code{active_speed} and \code{activity_level}}
  \item{\code{trap_rate}: number of camera trap records per unit time}
 }
 and columns
\itemize{
  \item{\code{estimate}: parameter estimates}
  \item{\code{se}: standard error}
  \item{\code{cv}: proportional coefficient of variation}
  \item{\code{lcl95}, \code{ucl95}: lower and upper 95\% confidence limits}
  \item{\code{n}: sample size}
  \item{\code{unit}: the unit of the estimate}
}
}
\description{
Creates a table of REM parameters taken from models for detection radius,
detection angle, speed and activity level.
}
\examples{
  \dontrun{
    pkg <- read_camtrapDP("./datapackage/datapackage.json")
    }
  data(pkg)
  sp <- "Vulpes vulpes"
  trdata <- get_traprate_data(pkg, species=sp)
  radmod <- fit_detmodel(radius~1, pkg, species=sp, order=0)
  angmod <- fit_detmodel(angle~1, pkg, species=sp, unit="radian", order=0)
  spdmod <- fit_speedmodel(pkg, species=sp)
  actmod <- fit_actmodel(pkg, species=sp, reps=100)
  get_parameter_table(trdata, radmod, angmod, spdmod, actmod)
}
