% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{get_rem_data}
\alias{get_rem_data}
\title{Get REM data from a camtrap-dp datapackage}
\usage{
get_rem_data(
  package,
  species = NULL,
  unit = c("day", "hour", "minute", "second")
)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{\link[camtraptor]{read_camtrap_dp}}.}

\item{species}{A character string indicating species subset to extract
data for; if NULL runs select_species to get user input.}

\item{unit}{The time unit in which to return camera effort.}
}
\value{
A tibble with columns:
  - \code{locationName}: name of the camera location
  - \code{effort}: the camera time for the location
  - \code{unit}: the effort time unit
  - \code{scientificName}: the scientific name of the species data extracted
  - \code{n}: the observation counts
  - \code{stratumID}: stratum identifier (only if this column is present in
    \code{package$data$deployments})
}
\description{
Extracts a data table of observation counts and effort for each
camera location in a camtrap-dp data package.
}
\examples{
  pkg <- camtraptor::read_camtrap_dp("./data/datapackage.json")
  remdata <- get_rem_data(pkg)
}
