% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{get_trap_rate}
\alias{get_trap_rate}
\title{Get average trap rate from REM data}
\usage{
get_trap_rate(traprate_data, strata = NULL, reps = 999, seed = 42)
}
\arguments{
\item{traprate_data}{A dataframe containing (at least) columns \code{n} and
\code{effort}, as returned by \code{\link{get_traprate_data}}; if
\code{strata} supplied for stratified calculation, must also have column
\code{stratumID}.}

\item{strata}{A dataframe with one row per stratum, and columns
\code{stratumID} and \code{area}.}

\item{reps}{The number of bootstrap replicates to run.}

\item{seed}{Integer seed for reproducible bootstrap; `NULL` to skip setting a seed. Default 42.}
}
\value{
A dataframe with columns:
  - \code{estimate}: average trap rate
  - \code{se}: standard error
  - \code{cv}: proportional coefficient of variation
  - \code{lcl95}, \code{ucl95}: lower and upper 95% confidence limits
  - \code{n}: sample size (number of locations)
  - \code{unit}: the unit of the estimate
}
\description{
Calculates average trap rate and its bootstrapped error from a table of
per-location observation counts and camera time.
}
\examples{
  \dontrun{
    pkg <- read_camtrapDP("./datapackage/datapackage.json")
    }
  data(pkg)
  trdata <- get_traprate_data(pkg, species="Vulpes vulpes")
  get_trap_rate(trdata)
}
