% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{lnorm_confint}
\alias{lnorm_confint}
\title{Log-normal confidence interval}
\usage{
lnorm_confint(estimate, se, percent = 95)
}
\arguments{
\item{estimate}{Numeric estimate value(s)}

\item{se}{Standard error(s) of the estimate}
}
\value{
A dataframe with a row per estimate input, and columns \code{lcl}
  and \code{ucl} (lower and upper confidence limits).
}
\description{
Calculates approximate log-normal confidence intervals given an estimate
and its standard error.
}
\examples{
  lnorm_confint(10.13, 3.57)
}
