% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{map_deployments}
\alias{map_deployments}
\title{Plot a map of deployments}
\usage{
map_deployments(pkg, basemap = c("street", "satellite"), ...)
}
\arguments{
\item{basemap}{Basemap to plot, street (default) or satellite}

\item{...}{Additional arguments passed to
\code{\link[leaflet]{addCircleMarkers}}}

\item{package}{Camera trap data package object, as returned by
\code{\link[camtraptor]{read_camtrap_dp}}.}
}
\description{
Creates an OpenStreetMap street or satellite map over-plotted with
deployment locations.
}
\examples{
  \dontrun{pkg <- camtraptor::read_camtrapDP("./data/datapackage.json")}
  data(pkg)
  map_deployments(pkg)
}
