% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{map_traprates}
\alias{map_traprates}
\title{Plot a map of deployment trap rates}
\usage{
map_traprates(
  pkg,
  species = NULL,
  basemap = c("street", "satellite"),
  maxSize = 25,
  minSize = 3
)
}
\arguments{
\item{species}{A character string indicating species subset to analyse.
Use scientific names. If NULL runs select_species to get user input;
if all uses all data.}

\item{basemap}{Basemap to plot, street (default) or satellite}

\item{maxSize}{Maximum point size to plot.}

\item{minSize}{Minimum point size to plot.}

\item{package}{Camera trap data package object, as returned by
\code{\link[camtraptor]{read_camtrap_dp}}.}
}
\description{
Creates an OpenStreetMap street or satellite map over-plotted with
deployment locations, with points sized in proportion to trap rate for a
given species.
}
\examples{
  \dontrun{pkg <- camtraptor::read_camtrapDP("./data/datapackage.json")}
  data(pkg)
  map_traprate(pkg, species="Vulpes vulpes")
}
