% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{read_camtrap_dp_csv}
\alias{read_camtrap_dp_csv}
\title{Create a datapackage from csv files}
\usage{
read_camtrap_dp_csv(
  folder,
  tryFormats = c("\%Y-\%m-\%d \%H:\%M:\%OS", "\%Y/\%m/\%d \%H:\%M:\%OS",
    "\%Y:\%m:\%d \%H:\%M:\%OS")
)
}
\arguments{
\item{folder}{A character string giving the folder in which the csv files sit.}

\item{tryFormats}{A character string defining date-time format, passed to as.POSIXct.}
}
\value{
As for \code{\link[camtraptor]{read_camtrap_dp}} but with reduced
 package metadata.
}
\description{
Reads csv data from a folder into a list that functions like a camtrap-dp
object for the purposes of density estimation.
}
\details{
The folder must contain three csv files: deployments.csv, media.csv
 and observations.csv, each of which must contain at least the following fields:
 \itemize{
  \item{deployments: }{deploymentID, locationID, locationName, longitude, latitude,start,end}
  \item{media: }{mediaID, deploymentID, sequenceID, timestamp, filePath, fileName}
  \item{observations: }{observationID, deploymentID, sequenceID, mediaID, timestamp, scientificName, count, speed, radius, angle}
}
}
\examples{
  \dontrun{pkg <- camtraptor::read_camtrap_dp_csv("./data")}
}
