% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{read_camtrap_dp_V1}
\alias{read_camtrap_dp_V1}
\title{Read Camtrap DP V1.0 data}
\usage{
read_camtrap_dp_V1(file)
}
\arguments{
\item{file}{Path to a datapackage.json file.}
}
\value{
As for \code{\link[camtraptor]{read_camtrap_dp}}.
}
\description{
Reads data from a Camtrap DP V1.0 datapackage Agouti export, effectively
down-versioning to V0.1, and recalculating speeds to include observations
with zero time difference, and ensuring positions are correctly ordered.
}
\examples{
  \dontrun{pkg <- camtraptor::read_camtrap_dp_csv("./data/datapackage.json")}
}
