% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{recalc_events}
\alias{recalc_events}
\title{Recalculate event observations}
\usage{
recalc_events(pkg)
}
\arguments{
\item{pkg}{Camera trap data package list, as returned by
\code{\link{read_camtrapDP}}.}
}
\value{
A duplicate of the original package with event observation values
 recalculated.
}
\description{
Recalculates event observation values (individualPositionRadius,
individualPositionAngle, individualSpeed) from media observations.
}
\details{
The process requires \code{media} data and fails if this is not
 available in the datapackage. The datapackage should also contain
 positions data, but the original package is simply returned unmodified if
 the \code{positions} data table contains no data. In recalculating speeds,
 observations with zero time difference (hence infinite speed) impute time
 elapsed from average frame rate.
}
\examples{
  \dontrun{pkg2 <- recalc_events(pkg2)}
}
