% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{rem}
\alias{rem}
\title{Fit a random encounter model}
\usage{
rem(parameters)
}
\arguments{
\item{parameters}{A dataframe containing REM parameter estimates with
(at least) rows:
\itemize{
  \item{\code{radius}: effective detection radius}
  \item{\code{angle}: effective detection angle}
  \item{\code{overall_speed}: average animal speed (day range)}
  \item{\code{trap_rate}: animal observations per unit time}
}
and columns:
\itemize{
  \item{\code{estimate}: numeric parameter estimate}
  \item{\code{se}: numeric parameter standard error}
  \item{\code{unit}: character parameter units (see \code{\link{convert_units}}
    for allowable values)}
}}
}
\value{
A dataframe with the input parameters plus estimated density and
 its errors.
}
\description{
Estimates REM density given a dataframe of parameters and their errors.
}
\examples{
  pkg <- camtraptor::read_camtrap_dp("./data/datapackage.json")
  sp <- "Vulpes vulpes"
  radmod <- fit_detmodel(radius~1, pkg, species=sp, order=0)
  angmod <- fit_detmodel(angle~1, pkg, species=sp, unit="radian", order=0)
  spdmod <- fit_speedmodel(pkg, species=sp)
  actmod <- fit_actmodel(pkg, species=sp, reps=100)
  param <- get_parameter_table(pkg, radmod, angmod, spdmod, actmod, sp)
  rem(param)
}
