% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{subset_deployments}
\alias{subset_deployments}
\title{Subset a camera trap datapackage by deployment (deprecated)}
\usage{
subset_deployments(package, choice)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{\link[camtraptor]{read_camtrap_dp}}.}

\item{choice}{A logical expression using column names from the
deployments table.}
}
\value{
As for \code{\link[camtraptor]{read_camtrap_dp}}, with all
 data tables reduced according to the choice criteria at the deployment
 level.
}
\description{
Select a subset of deployments from a datapackage defined by
a choice based on columns in the deployments table. Functionality
now provided in filter_camtrap_dp.
}
\examples{
# subset excluding a location and including only October 2017
  \dontrun{
    pkg <- camtraptor::read_camtrapDP("./datapackage/datapackage.json")
    }
  data(pkg)
  subpkg <- subset_deployments(pkg,
                               locationName != "S01" &
                               start >= as.POSIXct("2017-10-01", tz="UTC") &
                               end <= as.POSIXct("2017-10-31", tz="UTC"))
}
