% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{write_rem_csv}
\alias{write_rem_csv}
\title{Write REM results to csv file}
\usage{
write_rem_csv(...)
}
\arguments{
\item{...}{One or more REM analysis objects, separated by commas.}
}
\value{
None - creates a csv file.
}
\description{
Writes one or more REM estimate tables to a single csv file, with
identifying columns added for project, datapackage, project dates and
species. Input must be REM analysis object(s) created using
\code{\link{rem_estimatel}}. The resulting file name is taken from the
project and current date, and the file is saved to the working directory.
}
\examples{
 \dontrun{
   foxREM <- rem_estimate(pkg_checked, check_deployments=FALSE, species="Vulpes vulpes")
   hhogREM <- rem_estimate(pkg_checked, check_deployments=FALSE, species="Erinaceus europaeus")
   write_rem_csv(foxREM, hhogREM)
   }
}
