% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{correct_time}
\alias{correct_time}
\title{Correct times for a given deployment in a camera trap datapackage}
\usage{
correct_time(package, depID = NULL, locName = NULL, wrongTime, rightTime)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{\link[camtraptor]{read_camtrap_dp}}.}

\item{depID}{A character value giving the deployment ID,
to be matched in package$data$deployments$deploymentID.}

\item{locName}{A character value giving the location name,
to be matched in package$data$deployments$locationName.}

\item{wrongTime}{A character or POSIX reference date-time recorded wrongly
by the camera.}

\item{rightTime}{A character or POSIX value giving the correct date-time
when the reference time was recorded.}
}
\value{
As for \code{\link[camtraptor]{read_camtrap_dp}}, with all
 date-times corrected by the difference between rightTime and wrongTime.
}
\description{
When a camera trap starts with the wrong time stamp, times
in all datapackage tables can be corrected given a reference time
recorded by the camera, the correct time for this reference time
and the ID of the deployment to correct.
}
\details{
One, but not both, of depID and locName must be provided, as single
 text values. If locName is provided, the deployment associated with this
 in pkg$data$deployments is corrected, but if locName is associated with more
 than one deployment the function does not run.
}
\examples{
\dontrun{
  pkg <- camtraptor::read_camtrap_dp("./datapackage/datapackage.json")
}
data(pkg)
pkg_corrected <- correct_time(pkg,
                              locName = "S01",
                              wrongTime = "2017-10-02 08:06:43",
                              rightTime = "2017-09-01 10:36:00")
}
