% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{fit_actmodel}
\alias{fit_actmodel}
\title{Fit an activity model}
\usage{
fit_actmodel(
  package,
  species = NULL,
  reps = 999,
  obsdef = c("individual", "sequence"),
  ...
)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{\link[camtraptor]{read_camtrap_dp}}.}

\item{species}{A character string indicating species subset to analyse;
if NULL runs select_species to get user input.}

\item{reps}{Number of bootstrap replicates to run.}

\item{obsdef}{Observation definition, either individual or sequence.}

\item{...}{Arguments passed to \code{\link[activity]{fitact}}.}
}
\value{
An `actmod` list.
}
\description{
Fits an activity model to data package data and estimates activity
level (proportion of time spent active).
}
\examples{
  \dontrun{
    pkg <- camtraptor::read_camtrap_dp("./datapackage/datapackage.json")
  }
  data(pkg)
  \dontrun{
    # With interactive species definition
    act_model <- fit_actmodel(pkg)
  }
  # With species predefined, reps reduced for speed
  act_model <- fit_actmodel(pkg, species="Vulpes vulpes", reps=100)
  act_model@act
}
\seealso{
\code{\link[activity]{fitact}}
}
