% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{fit_speedmodel}
\alias{fit_speedmodel}
\title{Estimate average animal speed}
\usage{
fit_speedmodel(
  package,
  species = NULL,
  distUnit = c("m", "km", "cm"),
  timeUnit = c("second", "minute", "hour", "day")
)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{\link[camtraptor]{read_camtrap_dp}}.}

\item{species}{A character string indicating species subset to analyse;
if NULL runs select_species to get user input.}
}
\value{
List with elements:
 - \code{speed}: a one row dataframe containing columns \code{estimate}
   (mean) and \code{se} (standard error) speed while active.
 - \code{data}: a numeric vector of the data from which the estimate is derived
}
\description{
Calculates harmonic mean and standard error of animal speed while active
from a data package.
}
\examples{
  pkg <- camtraptor::read_camtrap_dp("./data/datapackage.json")
  speed_model <- fit_speedmodel(pkg)
  speed_model$speed
}
