% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{get_multiplier}
\alias{get_multiplier}
\title{Get a unit multiplier}
\usage{
get_multiplier(unitIN, unitOUT)
}
\arguments{
\item{unitIN}{A character vector giving the units of input.}

\item{unitOUT}{A character vector giving the units of output,
the same length as \code{unitIN},}
}
\value{
A vector of numbers giving the amount by which to multiply input
  values to arrive at unit-converted values.
}
\description{
Returns a multiplier to convert values from one unit to another, in
one of four categories: distance, time, angle, count.
}
\details{
Possible \code{unitIN} and \code{unitOUT} values are "cm", "m", "km"
 for distances; "second", "minute", "hour", "day", "100day" for times;
 "radian", "degree" for angles; "n" for count; "none" for no units. Unit
 ratios are allowed for rates or densities. In this case, units should be
 separated with a forward slash (e.g. "n/day", "km/hour", "n/km2"). Input
 and output types must match.
}
\examples{
  get_multiplier(c("m", "m/second"), c("km", "km/day"))
}
