% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{get_parameter_table}
\alias{get_parameter_table}
\title{Create a parameter table from a set of models}
\usage{
get_parameter_table(
  package,
  radius_model,
  angle_model,
  speed_model,
  activity_model,
  species = NULL,
  strata = NULL,
  reps = 999
)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{\link[camtraptor]{read_camtrap_dp}}.}

\item{radius_model}{A detection radius model fitted using \code{\link{fit_detmodel}}}

\item{angle_model}{A detection angle model fitted using \code{\link{fit_detmodel}}}

\item{speed_model}{A speed model fitted using \code{\link{fit_speedmodel}}}

\item{activity_model}{An activity  model fitted using \code{\link{fit_actmodel}}}

\item{species}{A character string indicating species subset to analyse;
if NULL runs select_species to get user input.}

\item{strata}{A dataframe of stratum information passed to \code{\link{get_trap_rate}}}

\item{reps}{Number of bootstrap replicates for estimating trap rate error
(see \code{\link{get_trap_rate}})}
}
\value{
A dataframe of unit-harmonised parameter estimates with seven rows:
  - \code{radius}: detection radius
  - \code{angle}: detection angle
  - \code{active_speed}: speed while active
  - \code{activity_level}: proportion of time spent active
  - \code{overall_speed}: long-term average speed (day range) - the product
    of \code{active_speed} and \code{activity_level}
  - \code{trap_rate}: number of camera trap records per unit time
 and seven columns:
  - \code{estimate}: parameter estimates
  - \code{se}: standard error
  - \code{cv}: proportional coefficient of variation
  - \code{lcl95}, \code{ucl95}: lower and upper 95% confidence limits
  - \code{n}: sample size
  - \code{unit}: the unit of the estimate
}
\description{
Creates a table of REM parameters taken from models for detection radius,
detection angle, speed and activity level.
}
\examples{
  pkg <- camtraptor::read_camtrap_dp("./data/datapackage.json")
  sp <- "Vulpes vulpes"
  radmod <- fit_detmodel(radius~1, pkg, species=sp, order=0)
  angmod <- fit_detmodel(angle~1, pkg, species=sp, unit="radian", order=0)
  spdmod <- fit_speedmodel(pkg, species=sp)
  actmod <- fit_actmodel(pkg, species=sp, reps=100)
  get_parameter_table(pkg, radmod, angmod, spdmod, actmod, sp)
}
