% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{plot_deployment_schedule}
\alias{plot_deployment_schedule}
\title{Plot a deployment Gantt chart}
\usage{
plot_deployment_schedule(package)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{\link[camtraptor]{read_camtrap_dp}}.}
}
\description{
Plots a Gantt chart illustrating deployment times (black lines) and
the occurrence of observations within those deployments (red bars). Useful
for checking errors in specification of deployment start and end dates, and
visualising spatiotemporal distribution of observations.
}
\examples{
  pkg <- camtraptor::read_camtrap_dp("./data/datapackage.json")
  plot_deployment_schedule(pkg)
}
