% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{read_camtrapDP}
\alias{read_camtrapDP}
\title{Read a Camptrap DP datapackage}
\usage{
read_camtrapDP(
  file,
  media = TRUE,
  recalc = TRUE,
  addV0 = TRUE,
  dtFormat = "YmdHMSz"
)
}
\arguments{
\item{file}{Path to a datapackage.json file.}

\item{media}{Logical defining whether to read media table.}

\item{recalc}{Logical defining whether to recalculate event; if \code{TRUE},
calls \code{\link{recalc_events}}.}

\item{addV0}{Logical defining whether to add camtrapDP version 0 fields, for
compatibility with functions using this data model.}

\item{dtFormat}{Character format used to read date-time fields, passed to
\code{\link[lubridate]{parse_date_time}}}
}
\value{
A list of data package metadata components and a \code{data}
 component with dataframes: \code{deployments}, \code{media},
 \code{observations} (containing event observations) and \code{positions}
 (containing media observations). If the \code{media} argument is FALSE,
 the \code{media} component is NULL. The \code{positions} may also be NULL
 if there are no media observations in the input.
}
\description{
Reads the metadata and csv files from a Camtrap DP version 1 datapackage
as exported from Agouti.
}
\examples{
  \dontrun{pkg <- read_camtrapDP("./data/datapackage.json")}
}
