% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{read_camtrapDP}
\alias{read_camtrapDP}
\title{Read Camtrap DP data}
\usage{
read_camtrapDP(file)
}
\arguments{
\item{file}{Path to a datapackage.json file.}
}
\value{
As for \code{\link[camtraptor]{read_camtrap_dp}}.
}
\description{
Reads data from a Camtrap DP datapackage Agouti export in either V0.1 or V1.
If data V0.1, uses camtraptor::read_camtrap_dp. If data are V1, effectively
down-versions to V0.1, and recalculates speeds. Positions are reordered
correctly before calculating speed (not always the case in raw output).
Observations with zero time difference (hence infinite speed) impute time
for speed calculation from average frame rate.
}
\examples{
  \dontrun{pkg <- read_camtrapDP("./data/datapackage.json")}
}
