% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{rem_estimate}
\alias{rem_estimate}
\title{Integrated random encounter model density estimate}
\usage{
rem_estimate(
  package,
  check_deployments = TRUE,
  species = NULL,
  radius_model = NULL,
  angle_model = NULL,
  speed_model = NULL,
  activity_model = NULL,
  strata = NULL,
  reps = 999
)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{\link[camtraptor]{read_camtrap_dp}}.}

\item{check_deployments}{Logical indicating whether to check deployment
calibration model diagnostic plots. If \code{TRUE} (default) runs
\code{check_deployment_models}; radius, angle and speed data from any
excluded deployments are then dropped from analysis. If \code{FALSE}
all data are used.}

\item{species}{A character string indicating species subset to analyse;
if NULL runs \code{\link{select_species}} to get user input.}

\item{radius_model}{A detection function model for radii fitted using
\code{\link{fit_detmodel}}.}

\item{angle_model}{A detection function model for angles fitted using
\code{\link{fit_detmodel}} with \code{unit} argument "radian" or "degree".}

\item{speed_model}{A speed model fitted using \code{\link{fit_speedmodel}}.}

\item{activity_model}{An activity model fitted using
\code{\link[activity]{fitact}} or \code{\link{fit_actmodel}}.}

\item{strata}{A dataframe of stratum areas, passed to \code{\link{get_trap_rate}}.}

\item{reps}{Number of bootstrap replicates for error estimation.}
}
\value{
A dataframe containing estimates and their errors for density and
  all contributing parameters.
}
\description{
Estimates animal density for a given species given a camtrapDP datapackage.
Models for detection radius and angle, speed and/or activity level can be
fitted externally and provided as arguments, or are fitted internally if not
provided (\code{NULL} default). Input units are assumed to be distance in m
and time in seconds.
}
\examples{
  # Load data
  \dontrun{
    pkg <- camtraptor::read_camtrap_dp("./datapackage/datapackage.json")
    }
  data(pkg)
  # Sense check deployment schedules
  plot_deployment_schedule(pkg)
 \dontrun{
   # Sense check deployment calibration model diagnostic plots
   pkg_checked <- check_deployment_models(pkg)
   # Fully automated analysis (interactive species definition by default;
   # reps reduced to limit run time).
   res <- rem_estimate(pkg_checked, check_deployments=FALSE, reps=100)
  }
  # Automated analysis with species predefined and no deployment checking
  sp <- "Vulpes vulpes"
  res <- rem_estimate(pkg, species=sp, check_deployments=FALSE, reps=100)
  # Inspect results
  res$estimates

  # Analysis with radius model fitted separately
  radmod <- fit_detmodel(radius~1, pkg, species=sp, truncation=15, order=0)
  res <- rem_estimate(pkg, check_deployments = FALSE, species = sp,
                      radius_model = radmod, reps=100)
  res$estimates
}
