% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{select_species}
\alias{select_species}
\title{Select a species name}
\usage{
select_species(package)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{\link[camtraptor]{read_camtrap_dp}}.}
}
\value{
A character string, scientific species name.
}
\description{
Presents a table of species names with observation count for each
and allows the user to interactively select one.
}
\examples{
  pkg <- camtraptor::read_camtrap_dp("./data/datapackage.json")
  species <- select_species(pkg)
}
