% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{select_species}
\alias{select_species}
\title{Select a species name}
\usage{
select_species(package, species = NULL)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{\link[camtraptor]{read_camtrap_dp}}.}

\item{species}{NULL (default) to select interactively, or character vector
giving either one or more valid species names found within
\code{package$data$observations$scientificName}, or the string "all"
to return all available species names.}
}
\value{
A character vector of one or more scientific species names.
}
\description{
Presents a table of species names with observation count for each
and allows the user to interactively select one.
}
\examples{
  \dontrun{
    pkg <- read_camtrapDP("./data/datapackage.json")
    select_species(pkg)
    # If provided, a valid species name is simply passed through
    select_species(pkg, "Vulpes vulpes")
    # Providing a species name that isn't found in the data throws an error
    select_species(pkg, "Vulpes vuppes")
  }
}
