% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camtrapDensity.R
\name{slice_camtrap_dp}
\alias{slice_camtrap_dp}
\title{Take a time slice of a data package}
\usage{
slice_camtrap_dp(package, start = NULL, end = NULL, depChoice = NULL)
}
\arguments{
\item{package}{Camera trap data package object, as returned by
\code{\link[camtraptor]{read_camtrap_dp}}.}

\item{start, end}{Single character or POSIXct values defining the
time range within which to slice the package.}

\item{depChoice}{A logical expression using column names from the
deployments table defining which deployments to slice.}
}
\value{
As for \code{\link[camtraptor]{read_camtrap_dp}}, with all
 data tables reduced according to the choice criteria.
}
\description{
Discards any observations, media, or deployments that fall wholly outside
the time range defined by start and end. When start or end are not
specified, no slicing is applied to start or end points respectively.
Slicing applies to all deployments by default, or can be applied to only a
subset of deployments specified by depChoice. The fully default behaviour
for this function is therefore to do nothing (the datapackage is returned
unchanged).
}
\examples{
  \dontrun{
    pkg <- read_camtrapDP("./datapackage/datapackage.json")
    }
  data(pkg)
# Slicing the whole package to mid October 2017
  subpkg <- slice_camtrap_dp(pkg,
                              start = "2017/10/10",
                              end = "2017/10/20")
# Slicing only deployments at location "S03" to a specific start time/date
  subpkg <- slice_camtrap_dp(pkg,
                              start = "2017/10/15 14:30:00",
                              depChoice = locationName=="S03")
}
